p2p: netPipe for <Go1.10 in own file with own build tag
Follow up of 283544c7f3ada24378fd6e0ac2f82cd927bc95e6
putting <Go1.10 implementation of netPipe in its own
file and protect it with its separate build tag.
